+++ app/Http/Controllers/Admin/Bunpou/BunpouModuleQuestionController.php
@@ -7,6 +7,7 @@ use App\Models\BunpouModuleQuestion;
 use App\Models\BunpouModuleTest;
 use App\Models\BunpouModuleAnswer;
 use App\Models\BunpouModules;
+use App\Models\BunpouModuleResult;
 use App\Repositories\BunpouModuleQuestionRepository;
 use App\Repositories\BunpouModuleAnswerRepository;
 use App\Http\Controllers\Controller;
@@ -15,7 +16,7 @@ use App\DataTables\BunpouModuleQuestionDatatable;
 class BunpouModuleQuestionController extends Controller
 {
 
-    protected $question, $repoQuestion, $test, $answer, $repoAnswer, $module;
+    protected $question, $repoQuestion, $test, $answer, $repoAnswer, $module, $result;
 
     public function __construct()
     {
@@ -23,6 +24,7 @@ class BunpouModuleQuestionController extends Controller
         $this->test = new BunpouModuleTest();
         $this->answer = new BunpouModuleAnswer();
         $this->module = new BunpouModules();
+        $this->result = new BunpouModuleResult();
         $this->repoQuestion = new BunpouModuleQuestionRepository();
         $this->repoAnswer = new BunpouModuleAnswerRepository();
     }
@@ -35,28 +37,13 @@ class BunpouModuleQuestionController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index(BunpouModuleQuestionDatatable $dataTable, $test=null)
+    public function index(BunpouModuleQuestionDatatable $dataTable)
     {
-        if($test==null){
-            return redirect()->route("bunpou.module.index");
-        }
-
-        if(request()->module!=null){
-            $data = $this->test->data([
-                ["module", "=", request()->module],
-                ["is_active", "=", true],
-            ])->orderBy("order","asc")->firstOrFail();
-            return redirect()->route($this->redirectAfterSave,$data->id);
-        }
-
-        $data = $this->test->data($test)->firstOrFail();
-        $test = $this->test->data([
-            ["is_active", "=", true],
-            ["module", "=", $data->module],
-        ])->get();
         $modules = $this->module->withTestCount();
-
-        return $dataTable->render('backend.bunpou.module.question.index',compact("data","test","modules"));
+        $data = request()->module!=null
+            ? $this->module->data(request()->module)->firstOrFail()
+            : $modules[0];
+        return $dataTable->render('backend.bunpou.module.question.index',compact("data","modules"));
     }
 
     /**
@@ -64,14 +51,10 @@ class BunpouModuleQuestionController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create($test=null)
+    public function create()
     {
-        if($test==null){
-            return redirect()->route("bunpou.module.index");
-        }
-
-        $test = $this->test->data($test)->firstOrFail();
-        return view('backend.bunpou.module.question.form',compact("test"));
+        $modules = $this->module->withTestCount();
+        return view('backend.bunpou.module.question.form',compact("modules"));
     }
 
     /**
@@ -82,13 +65,13 @@ class BunpouModuleQuestionController extends Controller
      */
     public function store(BunpouModuleQuestionRequest $request)
     {
-        $test = $this->test->data($request->test)->firstOrFail();
+        $this->test->data($request->test)->firstOrFail();
         $param = $request->all();
 
         $question = $this->repoQuestion->create($param);
         if (empty($question)) {
             flashDataAfterSave($question,$this->moduleName);
-            return redirect()->route($this->redirectAfterSave,$test->id);
+            return redirect()->route($this->redirectAfterSave);
         }
 
         $param["question"] = $question->id;
@@ -98,7 +81,7 @@ class BunpouModuleQuestionController extends Controller
         }
 
         flashDataAfterSave($answer,$this->moduleName);
-        return redirect()->route($this->redirectAfterSave,$test->id);
+        return redirect()->route($this->redirectAfterSave);
 
     }
 
@@ -124,12 +107,12 @@ class BunpouModuleQuestionController extends Controller
      */
     public function edit($id)
     {
-
         $question = $this->question->data($id)->firstOrFail();
         $answers = $this->answer->data("question",$id)->orderBy("order","asc")->get();
         $test = $this->test->data($question->test)->firstOrFail();
+        $modules = $this->module->withTestCount();
 
-        return view('backend.bunpou.module.question.form', compact('question','test','answers'));
+        return view('backend.bunpou.module.question.form', compact('question','test','answers','modules'));
     }
 
     /**
@@ -141,19 +124,19 @@ class BunpouModuleQuestionController extends Controller
      */
     public function update(BunpouModuleQuestionRequest $request, $id)
     {
-        $test = $this->test->data($request->test)->firstOrFail();
+        $this->test->data($request->test)->firstOrFail();
         $param = $request->all();
 
         $question = $this->repoQuestion->update($param, $id);
         if (empty($question)) {
             flashDataAfterSave($question,$this->moduleName);
-            return redirect()->route($this->redirectAfterSave,$test->id);
+            return redirect()->route($this->redirectAfterSave);
         }
 
         $answer = $this->repoAnswer->update($param, $id);
 
         flashDataAfterSave($answer,$this->moduleName);
-        return redirect()->route($this->redirectAfterSave,$test->id);
+        return redirect()->route($this->redirectAfterSave);
     }
 
     /**
@@ -164,6 +147,13 @@ class BunpouModuleQuestionController extends Controller
      */
     public function destroy($id)
     {
+        $question = $this->question->data($id)->firstOrFail();
+
+        $result = $this->result->data("test",$question->test)->first();
+        if($result!=null){
+            return "false";
+        }
+
         $this->repoAnswer->delete($id);
         return $this->repoQuestion->delete($id);
     }
@@ -176,10 +166,5 @@ class BunpouModuleQuestionController extends Controller
     public function deactivate($id)
     {
         return $this->repository->deactivate($id);
-    }
-
-    public function redirect(){
-        $test = $this->test->isActive()->firstOrFail();
-        return redirect()->route($this->redirectAfterSave,$test->id);
     }
 }
